Gate API tests with required features
